requirements.txt
backports.tarfile==1.2.0
celery==5.4.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-modeltranslation==0.19.11
django-redis==5.4.0
django-unfold==0.42.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
inflect==7.3.1
jaraco.collections==5.1.0
packaging==24.2
pip-chill==1.0.3
platformdirs==4.3.6
psycopg2-binary==2.9.10
tomli==2.2.1
uvicorn==0.32.1
django-ckeditor-5==0.2.15
channels==4.2.0
django-rosetta==0.10.1
django-cacheops~=7.1
django-silk
jst-django-core~=1.1.0
rich
pydantic
bcrypt
faker~=33.1.0
GDAL==3.9.3

# !NOTE: on-server
# gunicorn


django-storages
boto3


# !NOTE: on-websocket
# websockets
# channels-redis
